move mavenCentral() gradle repo in order to give custom repository configurations priority
In order to allow custom gradle dependency resolution to work, the buildscripts need to have the customizable repositories at the first place inside the repository decleration. The reason is that currently it is not possible to overwrite the repositories inside buildscripts of script plugins which are integrated with apply. If the customizable maven repository is not on top then you cant build the project without internet access.